Update main file tu use JS modules
